Feature request InfiniteAmmoInc#70 resizing nodes  * Allows node resizing by dragging the bottom right corner of a node.    It forces it's shape to be a square. I've added this limitation    because it would require too much work on the maths of the connection    arrows otherwise.  * Save to file in json, xml and yarn format. Twine doesn't handle    resizing, so the size is not saved in the twee and tw2 format.  * Loading a file also handle the size if it finds it in json, xml and    yarn format.  * Had to change a little the math of the arrows because the size was    hardcoded.
I've tried opening saved files containing the new 'size' option, and
old versions of Yarn (the one linked in the build section of the readme)
handled it correctly, ignoring the 'size' option and parsing the rest of
the file correctly.
And old saved files without the new 'size' option are also read
correctly, the 'size' option is optional when loading.

Note: the dragging box was already here in the html, but it was commented
out because it was incomplete. The author who added it was planning to
add a 'toggle expand' feature but apparently never did. I reused his
'resize' box and removed the few lines of code related to the
'toggle expand' feature.